+++ NEWS.md
@@ -4,6 +4,7 @@
 * Error messages displayed to the user are more informative on the actual error.
 * `filter_data_server()`: new argument `value_na` to set default value for NA's filters widgets.
 * `import_copypaste_ui()`: new argument `name_field` to show or not name field.
+* `import_copypaste_server()`: new argument `fread_args` to pass arguments to `data.table::fread`.
 * i18n: chinese translations added, thanks to [@xmusphlkg](https://github.com/xmusphlkg).
 * i18n: spanish translations added, thanks to [@drdsdaniel](https://github.com/drdsdaniel).
 * i18n: german translations added, thanks to [@SteEcker](https://github.com/SteEcker) and [@joerghenkebuero](https://github.com/joerghenkebuero).